I renamed the checkbox and ameliorated default FAB margin behaviour
